new page + income input & update works
added a new HTML page to take initial input from the user for their income. it can later be updated in the profile page and saved into the database.

+ new income.html
+ auth.py changes
+ modesl.py changes
+ views.py changes

changes needed to be made
++ styling on quiz page
link to income page needs to fit the styling of the page
++ general styling
need to update the style of some pages